requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
reportlab==4.0.9
python-dotenv==1.0.1
flask-cors==4.0.0
flask-limiter==3.5.0
waitress==3.0.0
gunicorn==22.0.0
stripe==7.10.0
